+++ core/middleware.py
@@ -29,15 +29,28 @@ class UserSpecificRedirectMiddleware(MiddlewareMixin):
 
 
 class StoreUserExpertiseMiddleware(MiddlewareMixin):
-    def process_request(self, request):
+
+    def should_set_product_expertise(self, request):
         if request.user.is_anonymous or 'remember-expertise-products-services' not in request.GET:
-            return
+            return False
 
+        if not request.user.company:
+            # no company yet. `update_company_profile` will update or create if not yet exists.
+            return True
+
+        # only update if specified products are different to current expertise
         products = request.GET.getlist('product')
+        return request.user.company and products and products != request.user.company.expertise_products_services
 
-        if request.user.company and products and products != request.user.company.expertise_products_services:
+    def process_request(self, request):
+        if self.should_set_product_expertise(request):
+            products = request.GET.getlist('product')
             helpers.update_company_profile(
                 sso_session_id=request.user.session_id,
                 data={'expertise_products_services': {'other': products}}
             )
-            request.user.company.data['expertise_products_services']['other'] = products
+            # invalidating the cached property
+            try:
+                del request.user.company
+            except AttributeError:
+                pass