+++ tests/unit/core/test_rules.py
@@ -16,7 +16,7 @@ def test_match_product_in_querystring(rf):
 @pytest.mark.django_db
 def test_not_match_product_in_querystring(rf):
     rule = factories.MatchProductExpertiseFactory()
-    request = rf.get(f'/?product=bla')
+    request = rf.get('/?product=bla')
     request.user = AnonymousUser()
     assert rule.test_user(request=request) is False
 
@@ -24,7 +24,7 @@ def test_not_match_product_in_querystring(rf):
 @pytest.mark.django_db
 def test_match_product_in_expertise(rf, user, mock_get_company_profile):
     rule = factories.MatchProductExpertiseFactory()
-    request = rf.get(f'/')
+    request = rf.get('/')
     request.user = user
 
     mock_get_company_profile.return_value = {'expertise_products_services': {'other': [rule.product.name]}}
@@ -36,7 +36,7 @@ def test_match_product_in_expertise(rf, user, mock_get_company_profile):
 def test_not_match_product_in_expertise(rf, user, mock_get_company_profile):
     mock_get_company_profile.return_value = {'expertise_products_services': {'other': ['foo']}}
     rule = factories.MatchProductExpertiseFactory()
-    request = rf.get(f'/')
+    request = rf.get('/')
     request.user = user
 
     assert rule.test_user(request=request) is False
@@ -53,6 +53,15 @@ def test_match_product_in_querystring_takes_precedence_over_user_expertise(rf, m
 
 
 @pytest.mark.django_db
+def test_match_product_no_company(rf, user):
+    rule = factories.MatchProductExpertiseFactory()
+    request = rf.get(f'/?product={rule.product.name}')
+    request.user = user
+
+    assert rule.test_user(request=request) is True
+
+
+@pytest.mark.django_db
 def test_match_country_in_querystring(rf):
     rule = factories.MatchCountryQuerystringFactory()
     request = rf.get(f'/?country={rule.country.name}')
@@ -63,7 +72,7 @@ def test_match_country_in_querystring(rf):
 @pytest.mark.django_db
 def test_not_match_country_in_querystring(rf):
     rule = factories.MatchCountryQuerystringFactory()
-    request = rf.get(f'/?country=bla')
+    request = rf.get('/?country=bla')
     assert rule.test_user(request=request) is False